+++ app/src/scenes/phase1/scenes/affected/components/StepCard.jsx
@@ -8,7 +8,7 @@ export function StepCard({ state = "todo", stepNumber, children }) {
         <div className="translate-x-5 flex-none flex h-9 w-9 items-center justify-center rounded-full bg-emerald-500">
           <BsCheck2 className="h-5 w-5 text-white" />
         </div>
-        <div className="w-full bg-white rounded-xl border p-8">{children}</div>
+        <div className="w-full bg-white rounded-xl border p-8 md:px-10">{children}</div>
       </div>
     );
   }
@@ -17,7 +17,7 @@ export function StepCard({ state = "todo", stepNumber, children }) {
     return (
       <div className="w-full flex items-center -ml-4">
         <div className="translate-x-5 flex-none flex h-9 w-9 items-center justify-center rounded-full bg-white border">{stepNumber}</div>
-        <div className="w-full bg-white rounded-xl border p-8">{children}</div>
+        <div className="w-full bg-white rounded-xl border-blue-600 border-[1px] p-8 md:px-10">{children}</div>
       </div>
     );
   }
@@ -25,10 +25,8 @@ export function StepCard({ state = "todo", stepNumber, children }) {
   if (state === "disabled") {
     return (
       <div className="flex items-center -ml-4">
-        <div className="translate-x-5 flex-none flex h-9 w-9 items-center justify-center rounded-full border bg-white">
-          <BsCheck2 className="h-5 w-5 text-gray-500" />
-        </div>
-        <div className="w-full bg-white rounded-xl border p-8">{children}</div>
+        <div className="translate-x-5 flex-none flex h-9 w-9 items-center justify-center rounded-full bg-white border">{stepNumber}</div>
+        <div className="w-full bg-white rounded-xl border p-8 md:px-10">{children}</div>
       </div>
     );
   }
